Ensure that `isMutexLocked()` works against a fresh database.
Due to changes in the version of Groovy used by Grails, when a `null` is passed to
`isMutexLocked( DistributedMutex mutex)` it can not determine which overloaded method
to call. As a result, this caused a `GroovyRuntimeException` to be thrown when on a
fresh db as the `DistributedMutex.findByIdentifier()` returns `null`.