Correct support for nested transactions.
Makes sure that the locking of the mutex occurs in another transaction so that it is not
tied to the transaction it might be called from within (say, a Service method). This change
is required due to the changes in the way `withTransaction()` work since about Grails 2.2 or 2.3.
It involves specifying the actual Spring Transaction propagation method for creating a new
transaction - whereas it used to be sufficient to just do `withSession { withTransaction { .. } }`.

Due to this change, it also means that the changes in the integration test are not rolled back at
the end of each test (as they're now in another transaction). So added `cleanup:` to all the tests
to at least free up the shared mutex. However, another option could be to ensure all entries in the
DistributedMutex directory are deleted.

In making this change I first demonstrated the issue with the provided additional test, then
resolved the issue to have the test go green.